Refactor to add more packages
