+++ src/main/java/impl/storage/TaskList.java
@@ -1,4 +1,4 @@
-package impl;
+package impl.storage;
 
 import impl.interfaces.Task;
